+++ mlf_core/create/domains/mlf_core_template_struct.py
@@ -13,7 +13,8 @@ class MlfcoreTemplateStruct:
     project_slug_no_hyphen: str = ''  # Project name mlf-core uses for almost all further processing without hyphens
     template_version: str = ''  # Version of the provided mlf-core template
     template_handle: str = ''  # Handle of the specific template, indicating which template is currently used
-    github_username: str = ''  # Github username
+    github_username: str = ''  # Github username (in case of an organization repository, the organization name)
+    creator_github_username: str = ''  # Github username of the person, that created the project
     is_github_repo: bool = False  # Whether the user wants to create a GitHub repo automatically
     is_repo_private: bool = False  # Whether to create a private Github repository
     is_github_orga: bool = False  # Whether Github repository is part of an organization